+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Add the possibility to use "pay" and "waive" fee/fine permissions separately. Refs UIU-3109.
 * Use showing/hiding approach for fee/fine action buttons. Refs UIU-3156.
 * Support sorting for each column in the Reading Room Access accordion in User Profile Edit. Refs UIU-3132.
+* Add the 'Preferred email communication' field in the Contact information accordion of the user profile. Refs UIU-3152.
 
 ## [10.1.1](https://github.com/folio-org/ui-users/tree/v10.1.1) (2024-05-07)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v10.1.0...v10.1.1)